separating error policy from configuration parser
