+++ include/micm/configure/solver_config.hpp
@@ -19,49 +19,12 @@
 #include <micm/system/species.hpp>
 #include <micm/system/system.hpp>
 #include <micm/util/constants.hpp>
+#include <micm/util/error_policies.hpp>
 #include <nlohmann/json.hpp>
 #include <variant>
 
 namespace micm
 {
-  // JSON-Parsing error policies
-  template<class Object>
-  class ThrowPolicy
-  {
-    class Exception : public std::exception
-    {
-     public:
-      const char* msg_;
-
-     public:
-      Exception(const char* msg)
-          : msg_(msg)
-      {
-      }
-
-      virtual const char* what()
-      {
-        return msg_;
-      }
-    };
-
-   public:
-    Object OnError(std::string message)
-    {
-      throw Exception(message.c_str());
-    }
-  };
-
-  template<class Object>
-  class NoThrowPolicy
-  {
-   public:
-    Object OnError(std::string message)
-    {
-      std::cerr << message << std::endl;
-      return Object();
-    }
-  };
 
   // Solver parameters
   struct SolverParameters
